+++ src/main/java/com/jdb/recovery/logs/CompensationLog.java
@@ -30,7 +30,7 @@ public class CompensationLog extends LogRecord {
     }
 
     @Override
-    public int getPageId() {
+    public long getPageId() {
         return 0;
     }
