+++ src/main/java/com/jdb/recovery/logs/MasterLog.java
@@ -19,7 +19,7 @@ public class MasterLog extends LogRecord {
     }
 
     @Override
-    public int getPageId() {
+    public long getPageId() {
         return 0;
     }
